logo.py
from os import path
from PIL import Image
import numpy as np
import matplotlib.pyplot as plt
import random

from wordcloud import WordCloud, STOPWORDS, ImageColorGenerator



def grey_color_func(word, font_size, position, orientation, random_state=None,
                    **kwargs):
    return "hsl(0, 0%%, %d%%)" % random.randint(60, 100)

# Read the whole text.
text = open('tweets.txt',encoding="utf-8").read()

# read the mask / color image taken from
# http://jirkavinse.deviantart.com/art/quot-Real-Life-quot-Alice-282261010
alice_coloring = np.array(Image.open("logo.jpg"))
stopwords = set(STOPWORDS)
stopwords.add("https")

wc = WordCloud(max_words=1000, mask=alice_coloring, stopwords=stopwords, margin=10,
               random_state=1).generate(text)
# store default colored image
default_colors = wc.to_array()
plt.title("Custom colors")
plt.imshow(wc.recolor(color_func=grey_color_func, random_state=3),
           interpolation="bilinear")
wc.to_file("a_new_hope.png")
plt.axis("off")
plt.figure()
plt.title("Default colors")
plt.imshow(default_colors, interpolation="bilinear")
plt.axis("off")
plt.show()